Add Prefs

+++ source/IOStream.c
@@ -23,7 +23,7 @@ int32_t Rb_IOStream_print(const Rb_IOStream* stream, const char* fmt, ...){
 
     va_end(vl);
 
-    if(stream->api.write(stream->handle, str, strlen(str) + 1) != strlen(str) + 1){
+    if(stream->api.write(stream->handle, str, strlen(str) + 1) != (int32_t)strlen(str) + 1){
         free(str);
 
         return RB_ERROR;
